NAV: Move and rename RsTraitMethodImplLineMarkerProvider
Now that we also annotate "const" and "type" items in trait impls,
the old name is not accurate anymore.

Also move the file to the "lineMarkers" package.
